# docs/openapi/application.yaml
openapi: 3.0.3
info:
  title: Exchange Application HTTP API
  version: 1.0.0
  description: Реестр всех HTTP operations, используемый для проверки Swagger drift.
paths:
...
  /api/v1/instruments/{instrumentId}:
    parameters:
      - $ref: '#/components/parameters/InstrumentId'
    get:
      operationId: getInstrument
      responses:
        '200': { description: Instrument rules history }
  /api/v1/realtime/instruments:
    get:
      operationId: listRealtimeInstruments
      security: [{ ApiKeyAuth: [] }]
      responses:
        '200': { description: Locally cached Twelve Data instrument catalog }
        '400': { description: Invalid filters or pagination }
  /api/v1/realtime/instruments/{instrumentId}:
    parameters:
      - $ref: '#/components/parameters/InstrumentId'
    get:
      operationId: getRealtimeInstrument
      security: [{ ApiKeyAuth: [] }]
      responses:
        '200': { description: Realtime instrument metadata }
        '404': { description: Realtime instrument not found }
  /api/v1/realtime/market-status:
    get:
      operationId: getRealtimeMarketStatus
      security: [{ ApiKeyAuth: [] }]
      responses:
        '200': { description: Bounded provider and ingest status }
  /api/v1/realtime/quotes:
    get:
      operationId: listRealtimeQuotes
      security: [{ ApiKeyAuth: [] }]
      parameters:
        - { name: instrumentIds, in: query, required: true, schema: { type: string, maxLength: 12900 } }
      responses:
        '200': { description: 'Fresh, stale or unavailable local quote snapshots' }
        '400': { description: Invalid or oversized instrument set }
...
components:
  securitySchemes:
    ApiKeyAuth: { type: apiKey, in: header, name: x-api-key }
    SessionCookie: { type: apiKey, in: cookie, name: __Host-exchange_session }
  parameters:
...
    InstrumentId:
      name: instrumentId
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
...